fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bcrypt==3.2.0
cassandra-driver==3.25.0
cffi==1.15.0
click==8.1.7
colorama==0.4.6
cryptography==37.0.4
exceptiongroup==1.2.2
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.0.3
MarkupSafe==2.1.1
mkl-service==2.4.0
mkl_fft>=1.3.0
mkl-random==1.2.2
numpy==1.24.3
pandas==2.2.3
passlib==1.7.4
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing-extensions>=4.8.0
tzdata==2024.2
uvicorn==0.32.0
scikit-learn

